refactor: move .get spec to mockClient suite
This test does not require a socket server and can be run only with a mock client.
